Social-Sync/backend/requirements.txt
pandas
numpy
scikit-learn
spotipy
mysql-connector-python
python-dotenv
Flask
requests
fastapi
uvicorn
